Add files via upload
app.js is where the game logic should be coded.

style.css contains all css code for the entire web app and also the responsiveness code which needs to be fixed.

index.html is where the content for the game (under <!--Game--->) and the pop-up modals (under <!--Pages--->).

pages.js is where the logic for the pop-up modals should be coded. This needs to be rebuilt.